type casting to avoid printf warning

+++ lib/krb5/send_to_kdc.c
@@ -1246,9 +1246,9 @@ krb5_sendto_context(krb5_context context,
 		"krb5_sendto_context %s done: %d hosts %lu packets %lu wc: %ld.%06ld nr: %ld.%06ld kh: %ld.%06ld tid: %08x",
 		realm, ret,
 		ctx->stats.num_hosts, ctx->stats.sent_packets,
-		stop_time.tv_sec, stop_time.tv_usec,
-		ctx->stats.name_resolution.tv_sec, ctx->stats.name_resolution.tv_usec,
-		ctx->stats.krbhst.tv_sec, ctx->stats.krbhst.tv_usec, ctx->stid);
+		stop_time.tv_sec, (long)stop_time.tv_usec,
+		ctx->stats.name_resolution.tv_sec, (long)ctx->stats.name_resolution.tv_usec,
+		ctx->stats.krbhst.tv_sec, (long)ctx->stats.krbhst.tv_usec, ctx->stid);
 
 
     if (freectx)
